requirements.txt
ccxt
pandas
pandas_ta
python-dotenv
numpy
